Add jQuery 1.10.2
Add jQuery 1.10.2 which is compatible with more versions of Internet
Explorer unlike jQuery 2.0.3.

Use the uncompressed version as sails will deal with the compression
itself.

Update Grunfile.js to load jQuery first.

jQuery 2.0.3-min was added previously for the POC version of the app, so
the NOTICE file is already up-to-date.